requirements.txt
# ============================================================
# FinTraceQA — Complete Python Dependencies
# ============================================================
# Install with:  pip install -r requirements.txt
# Tested with:   Python 3.12.1

# ── Backend (FastAPI server) ─────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pyyaml>=6.0

# ── AI / LLM ─────────────────────────────────────────────────
openai>=1.0.0

# ── Knowledge Graph (Neo4j) ──────────────────────────────────
neo4j>=5.0.0

# ── Language Detection & NLP utilities ───────────────────────
langdetect>=1.0.9
networkx>=3.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ── Financial KG builder ─────────────────────────────────────
google-generativeai>=0.5.0
aiohttp>=3.8.0
requests>=2.28.0
sentence-transformers>=2.2.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
pytz>=2023.3
rich>=13.4.0
tqdm>=4.65.0
huggingface-hub>=0.16.0

# ── Translation Model (training) ─────────────────────────────
torch>=2.1.0
transformers>=4.36.0
datasets>=2.14.0
sentencepiece>=0.1.99
tokenizers>=0.14.0
accelerate>=0.24.0
wandb>=0.16.0
scipy>=1.11.0
